CSTester updated to latest version. Components requirements may include not only components names and versions, but also their functionality. jQuery++ removed because of new bugs, jquery.cookie plugin added for cookie functionality. SimpleImage plugin added. Photo gallery module added. More intelligent autoloader, can load classes with names that differ from component name, but within it's namespace. Some fixes and refactoring in system core. A lot of other fixes.
System version 0.302
modules:
Blogs 0.050
Comments 0.012
Cron 0.006
Feedback 0.002
HybridAuth 2.1.1-cs24
OAuth2 0.016
Photo_gallery 0.001
Plupload 1.5.7-cs7
Static_pages 0.029
plugins:
Admin_default_theme 0.004
Fotorama 4.2.1-cs2
SimpleImage 2.3-cs1
TinyMCE 4.0.1-cs6

+++ components/modules/Plupload/trigger/installed.php
@@ -32,6 +32,8 @@ Trigger::instance()->register(
 		while ($f = $cdb->fs($files)) {
 			$storage->unlink($f);
 		}
-		$storage->rmdir('Plupload');
+		if ($storage->is_dir('Plupload')) {
+			$storage->rmdir('Plupload');
+		}
 	}
 );
